Moved tests to correct folder. 34 / 54 tests work.
