+++ tests/cannon_vs_mg_recoil_behavior.test.js
@@ -1,10 +1,10 @@
 import { describe, it, expect } from "vitest";
-import { createTank } from "../../entities/tankFactory.js";
-import { weaponInputSystem } from "../../systems/weapons/weaponInputSystem.js";
-import { weaponRecoilSystem } from "../../systems/weapons/weaponRecoilSystem.js";
+import { createTank } from "../src/entities/tankFactory.js";
+import { weaponInputSystem } from "../src/systems/weapons/weaponInputSystem.js";
+import { weaponRecoilSystem } from "../src/systems/weapons/weaponRecoilSystem.js";
 
 function makeReg() {
-  const { createRegistry } = require("../../engine/registry.js");
+  const { createRegistry } = require("../src/engine/registry.js");
   return createRegistry();
 }
