+++ tests/hardpointMountSystem_offsets.test.js
@@ -1,10 +1,10 @@
 import { describe, it, expect } from "vitest";
-import { hardpointMountSystem } from "../hardpointMountSystem.js";
-import { createTransform } from "../../../components/transform.js";
-import { createHardpoints, createMount } from "../../../components/hardpoints.js";
+import { hardpointMountSystem } from "../src/systems/attachment/hardpointMountSystem.js";
+import { createTransform } from "../src/components/transform.js";
+import { createHardpoints, createMount } from "../src/components/hardpoints.js";
 
 function makeReg() {
-  const { createRegistry } = require("../../../engine/registry.js");
+  const { createRegistry } = require("../src/engine/registry.js");
   return createRegistry();
 }
