+++ tests/recoil_flow.test.js
@@ -1,12 +1,12 @@
 import { describe, it, expect } from "vitest";
-import { createTank } from "../../entities/tankFactory.js";
-import { hardpointMountSystem } from "../../systems/attachment/hardpointMountSystem.js";
-import { weaponInputSystem } from "../../systems/weapons/weaponInputSystem.js";
-import { weaponRecoilSystem } from "../../systems/weapons/weaponRecoilSystem.js";
-import { transformApplySystem } from "../../systems/motion/transformApplySystem.js";
+import { createTank } from "../src/entities/tankFactory.js";
+import { hardpointMountSystem } from "../src/systems/attachment/hardpointMountSystem.js";
+import { weaponInputSystem } from "../src/systems/weapons/weaponInputSystem.js";
+import { weaponRecoilSystem } from "../src/systems/weapons/weaponRecoilSystem.js";
+import { transformApplySystem } from "../src/systems/motion/transformApplySystem.js";
 
 function makeReg() {
-  const { createRegistry } = require("../../engine/registry.js");
+  const { createRegistry } = require("../src/engine/registry.js");
   return createRegistry();
 }
